+++ HandTrackerRenderer.py
@@ -65,7 +65,7 @@ class HandTrackerRenderer:
                 if self.show_rot_rect:
                     cv2.polylines(self.frame, [np.array(hand.rect_points)], True, (0,255,255), 2, cv2.LINE_AA)
                 if self.show_landmarks:
-                    lines = [np.array([hand.landmarks[point] for point in line]).astype(np.int) for line in LINES_HAND]
+                    lines = [np.array([hand.landmarks[point] for point in line]).astype(np.int32) for line in LINES_HAND]
                     if self.show_handedness == 3:
                         color = (0,255,0) if hand.handedness > 0.5 else (0,0,255)
                     else:
